+++ processor/src/test/resources/input/com/example/override_generics/GenericsBaseModel.java
@@ -1,0 +1,11 @@
+package com.example.override_generics;
+
+import org.realityforge.arez.annotations.Action;
+
+public class GenericsBaseModel<V>
+{
+  @Action
+  public void foo( final V v )
+  {
+  }
+}