+++ processor/src/test/resources/input/com/example/override_generics/GenericsModel.java
@@ -1,0 +1,16 @@
+package com.example.override_generics;
+
+import org.realityforge.arez.annotations.Action;
+import org.realityforge.arez.annotations.ArezComponent;
+
+@ArezComponent
+public class GenericsModel
+  extends GenericsMiddleModel<Integer>
+{
+  @Action
+  @Override
+  public void foo( final Integer v )
+  {
+    super.foo( v );
+  }
+}